Change Timestamp from Int to Long, Add timestamp to Entities

+++ src/main/java/ru/m2mcom/pondnotes/service/dto/ChemicalsDTO.java
@@ -23,9 +23,9 @@ public class ChemicalsDTO implements Serializable {
     @NotNull
     private Double tempVal;
 
-    private Integer timestamp;
+    private Long timestamp;
 
-    private Integer userId;
+    private Long userId;
 
     public Long getId() {
         return id;
@@ -62,18 +62,18 @@ public class ChemicalsDTO implements Serializable {
     public void setTempVal(Double tempVal) {
         this.tempVal = tempVal;
     }
-    public Integer getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(Integer timestamp) {
+    public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
