+++ src/test/java/ru/m2mcom/pondnotes/web/rest/TankResourceIntTest.java
@@ -51,8 +51,8 @@ public class TankResourceIntTest {
     private static final String DEFAULT_DESCRIPTION = "AAAAAAAAAA";
     private static final String UPDATED_DESCRIPTION = "BBBBBBBBBB";
 
-    private static final Integer DEFAULT_TIMESTAMP = 1;
-    private static final Integer UPDATED_TIMESTAMP = 2;
+    private static final Long DEFAULT_TIMESTAMP = 1L;
+    private static final Long UPDATED_TIMESTAMP = 2L;
 
     @Autowired
     private TankRepository tankRepository;
@@ -211,7 +211,7 @@ public class TankResourceIntTest {
             .andExpect(jsonPath("$.[*].tankName").value(hasItem(DEFAULT_TANK_NAME.toString())))
             .andExpect(jsonPath("$.[*].tankType").value(hasItem(DEFAULT_TANK_TYPE.toString())))
             .andExpect(jsonPath("$.[*].description").value(hasItem(DEFAULT_DESCRIPTION.toString())))
-            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP)));
+            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP.intValue())));
     }
 
     @Test
@@ -228,7 +228,7 @@ public class TankResourceIntTest {
             .andExpect(jsonPath("$.tankName").value(DEFAULT_TANK_NAME.toString()))
             .andExpect(jsonPath("$.tankType").value(DEFAULT_TANK_TYPE.toString()))
             .andExpect(jsonPath("$.description").value(DEFAULT_DESCRIPTION.toString()))
-            .andExpect(jsonPath("$.timestamp").value(DEFAULT_TIMESTAMP));
+            .andExpect(jsonPath("$.timestamp").value(DEFAULT_TIMESTAMP.intValue()));
     }
 
     @Test
@@ -331,7 +331,7 @@ public class TankResourceIntTest {
             .andExpect(jsonPath("$.[*].tankName").value(hasItem(DEFAULT_TANK_NAME.toString())))
             .andExpect(jsonPath("$.[*].tankType").value(hasItem(DEFAULT_TANK_TYPE.toString())))
             .andExpect(jsonPath("$.[*].description").value(hasItem(DEFAULT_DESCRIPTION.toString())))
-            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP)));
+            .andExpect(jsonPath("$.[*].timestamp").value(hasItem(DEFAULT_TIMESTAMP.intValue())));
     }
 
     @Test
